+++ sdk/identity/azure_identity/src/azure_pipelines_credential.rs
@@ -206,7 +206,7 @@ mod tests {
     use super::*;
     use crate::env::Env;
     use azure_core::{
-        http::{BufResponse, ClientOptions, TransportOptions},
+        http::{BufResponse, ClientOptions, Transport},
         Bytes,
     };
     use azure_core_test::http::MockHttpClient;
@@ -254,7 +254,7 @@ mod tests {
         let options = AzurePipelinesCredentialOptions {
             credential_options: ClientAssertionCredentialOptions {
                 client_options: ClientOptions {
-                    transport: Some(TransportOptions::new(Arc::new(mock_client))),
+                    transport: Some(Transport::new(Arc::new(mock_client))),
                     ..Default::default()
                 },
                 ..Default::default()
@@ -321,7 +321,7 @@ mod tests {
         let options = AzurePipelinesCredentialOptions {
             credential_options: ClientAssertionCredentialOptions {
                 client_options: ClientOptions {
-                    transport: Some(TransportOptions::new(Arc::new(mock_client))),
+                    transport: Some(Transport::new(Arc::new(mock_client))),
                     ..Default::default()
                 },
                 ..Default::default()