+++ sdk/identity/azure_identity/src/client_secret_credential.rs
@@ -160,7 +160,7 @@ mod tests {
     use crate::tests::*;
     use azure_core::{
         authority_hosts::AZURE_PUBLIC_CLOUD,
-        http::{headers::Headers, BufResponse, StatusCode, TransportOptions},
+        http::{headers::Headers, BufResponse, StatusCode, Transport},
         Bytes, Result,
     };
     use std::vec;
@@ -203,7 +203,7 @@ mod tests {
             FAKE_SECRET.into(),
             Some(ClientSecretCredentialOptions {
                 client_options: ClientOptions {
-                    transport: Some(TransportOptions::new(Arc::new(sts))),
+                    transport: Some(Transport::new(Arc::new(sts))),
                     ..Default::default()
                 },
                 ..Default::default()
@@ -246,7 +246,7 @@ mod tests {
             FAKE_SECRET.into(),
             Some(ClientSecretCredentialOptions {
                 client_options: ClientOptions {
-                    transport: Some(TransportOptions::new(Arc::new(sts))),
+                    transport: Some(Transport::new(Arc::new(sts))),
                     ..Default::default()
                 },
                 ..Default::default()